refactor(LottoPrinter,LottoMachine,Application): Extract 'LottoTicket' generation from 'LottoMachine'
- Extract 'LottoTicket' generation part from 'LottoMachine' to new class 'LottoPrinter' to meet the requirement of the example output, which is proper progression of program for user interface.
